Skip Tor test on CI
